Fix renderBalanced kernel: remove broken unused kernel and add fp16 feature output
The renderBalanced kernel added in the NHT commit referenced undefined constants
(GUTParameters::Tiling::VirtualTilesPerTile etc.) and an undefined function
(initializeRayPerPixel). It was never called from gutRenderer.cu and therefore
served no purpose. Removing it fixes the CUDA compilation failure.

Also applies fp16 feature output support across the CUDA pipeline:
- TRadianceDensityElem type (float or __half) based on FEATURE_OUTPUT_HALF
- gutRenderer.cuh/gutRenderer.h/gutRenderer.cu: use TRadianceDensityElem* for
  radiance+density buffer in renderForward/renderBackward signatures
- splatRaster.cpp: allocate rayRadianceDensity as fp16 when FEATURE_OUTPUT_HALF=1
- rayPayload.cuh: finalizeRay writes __half via __float2half when FEATURE_OUTPUT_HALF
- rayPayloadBackward.cuh: initializeBackwardRay reads __half via __half2float
- setup_3dgut.py: emit -DFEATURE_OUTPUT_HALF compile define from config
- configs/render/3dgut.yaml: feature_output_half: true enables fp16 by default
- threedgrut/utils/render.py: apply_feature_decoder converts fp16→float32 for decoder
- threedgut_tracer/tracer.py: pred_opacity cast to float32 after fp16 tensor output

+++ threedgut_tracer/tracer.py
@@ -328,9 +328,10 @@ class Tracer:
             )
 
             # pred_rgba is [..., RAY_FEATURE_DIM + 1]: features (or RGB) + density
+            # May be fp16 when FEATURE_OUTPUT_HALF=1; opacity is always converted to fp32.
             ray_feature_dim = gaussians.ray_feature_dim
             pred_rgb = pred_rgba[..., :ray_feature_dim].unsqueeze(0).contiguous()
-            pred_opacity = pred_rgba[..., ray_feature_dim:].unsqueeze(0).contiguous()
+            pred_opacity = pred_rgba[..., ray_feature_dim:].float().unsqueeze(0).contiguous()
             pred_dist = pred_dist.unsqueeze(0).contiguous()
             hits_count = hits_count.unsqueeze(0).contiguous()
